# src/main/java/com/dliriotech/tms/authservice/dto/AuthResponse.java
package com.dliriotech.tms.authservice.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;

@Builder
@Data
@AllArgsConstructor
public class AuthResponse {
    private String token;
    private String refreshToken;
}